apps/platform/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
asyncpg==0.28.0
# psycopg2-binary is only needed for migration scripts
# On Windows, install from: https://pypi.org/project/psycopg2-binary/#files
# Or use: pip install psycopg2-binary --only-binary :all:
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pandas==2.1.1
google-analytics-data==0.18.2
httpx>=0.24,<0.26
apscheduler==3.10.4
cryptography==41.0.7
pytz==2023.3.post1

# Authentication
pyjwt==2.8.0
supabase==2.3.4

# Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Exports
openpyxl==3.1.2

# Monitoring
sentry-sdk[fastapi]==1.40.0